huge modifications to src and library structure
Courtesty of cmake:
- lib has been renamed to libnumerics, which made more sense in a naming
point of view. Lib means nothing, at least libnumerics means something
- gauss_elim.cpp has been removed from src and put in libnumerics, where
it should live
- matrix.[ch]pp has moved to libnumerics as well, it now uses hpp for
its headers. This may need to be updated in multiple places, so be aware
of that.
- CMakeLists updated

I also moved the pdf to the directory of the actual homework instead of
outside it. It didn't make much sense leaving this file outside
anyway, so I moved it in here where it will clutter everything a lot less.